Moved!!!
Ok, this project was moved over from the repo over on BitBucket. Its
just going to be a clean re-commit to github without any of the old
logs.  Gonna copy pasta the last commit that I made like two days ago to
get a feel for what I have updated,
"So,  I decided to pick up this project again and optimize it so there
isn't any more lag. As of now, I have redone the background following a
recent unity live training session, so the background scrolling code is
in a pseudo copy pasta state until I get around to changing it and
making it randomize tiles again. The codebase will also move over to
github whenever I get around to putting the time into moving it."